I'm finally at a point where I can say I'm done 😀 It has 80% of the functionality I wanted; and I'm 95% sure it has 100% of the functionality that I need.
The script compiles, and can be run stand alone (albeit with VLC). It creates a folder called `assets` in the current script directory.

VLC is still used to mirror the screen, which still experiences the limited ~2Hz refresh rate.

There is a GUI to send windows from the private to the public screen.